+++ bindings/nostr-sdk-ffi/src/protocol/event/tag/kind.rs
@@ -70,6 +70,12 @@ pub enum TagKind {
     Amount,
     /// Lnurl (NIP57)
     Lnurl,
+    /// MLS Protocol Version (NIP104)
+    MlsProtocolVersion,
+    /// MLS Cipher Suite (NIP104)
+    MlsCiphersuite,
+    /// MLS Extensions (NIP104)
+    MlsExtensions,
     /// Name tag
     Name,
     /// Url
@@ -168,6 +174,9 @@ impl<'a> From<tag::TagKind<'a>> for TagKind {
             tag::TagKind::Request => Self::Request,
             tag::TagKind::Web => Self::Web,
             tag::TagKind::Word => Self::Word,
+            tag::TagKind::MlsProtocolVersion => Self::MlsProtocolVersion,
+            tag::TagKind::MlsCiphersuite => Self::MlsCiphersuite,
+            tag::TagKind::MlsExtensions => Self::MlsExtensions,
             tag::TagKind::Custom(unknown) => Self::Unknown {
                 unknown: unknown.to_string(),
             },
@@ -226,6 +235,9 @@ impl<'a> From<TagKind> for tag::TagKind<'a> {
             TagKind::Request => Self::Request,
             TagKind::Web => Self::Web,
             TagKind::Word => Self::Word,
+            TagKind::MlsProtocolVersion => Self::MlsProtocolVersion,
+            TagKind::MlsCiphersuite => Self::MlsCiphersuite,
+            TagKind::MlsExtensions => Self::MlsExtensions,
             TagKind::Unknown { unknown } => Self::Custom(Cow::Owned(unknown)),
         }
     }