+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 
 ### Added
 
+* nostr: add NIP104 tag and event kinds ([erskingardner])
 * nostr: add `SingleLetterTag::as_str` and `TagKind::as_str` ([Yuki Kishimoto])
 * nostr: add `Kind::Comment` ([reyamir])
 * nostr: add `EventBuilder::comment` ([reyamir])
@@ -733,6 +734,7 @@ added `nostrdb` storage backend, added NIP32 and completed NIP51 support and mor
 [w3irdrobot]: https://github.com/w3irdrobot
 [nanikamado]: https://github.com/nanikamado
 [rodant]: https://github.com/rodant
+[erskingardner]: https://github.com/erskingardner
 
 <!-- Tags -->
 [Unreleased]: https://github.com/rust-nostr/nostr/compare/v0.36.0...HEAD